Added `de/arrays` module
