+++ crates/bevy_reflect/src/serde/de/deserializer.rs
@@ -1,9 +1,8 @@
+use crate::serde::de::arrays::ArrayVisitor;
 use crate::serde::de::structs::StructVisitor;
 use crate::serde::de::tuple_structs::TupleStructVisitor;
 use crate::serde::de::tuples::TupleVisitor;
-use crate::serde::de::{
-    ArrayVisitor, EnumVisitor, ListVisitor, MapVisitor, OptionVisitor, SetVisitor,
-};
+use crate::serde::de::{EnumVisitor, ListVisitor, MapVisitor, OptionVisitor, SetVisitor};
 use crate::serde::TypeRegistrationDeserializer;
 use crate::{PartialReflect, ReflectDeserialize, TypeInfo, TypeRegistration, TypeRegistry};
 use core::fmt::Formatter;
@@ -280,10 +279,7 @@ impl<'a, 'de> DeserializeSeed<'de> for TypedReflectDeserializer<'a> {
             TypeInfo::Array(array_info) => {
                 let mut dynamic_array = deserializer.deserialize_tuple(
                     array_info.capacity(),
-                    ArrayVisitor {
-                        array_info,
-                        registry: self.registry,
-                    },
+                    ArrayVisitor::new(array_info, self.registry),
                 )?;
                 dynamic_array.set_represented_type(Some(self.registration.type_info()));
                 Ok(Box::new(dynamic_array))
